+++ mlx_lm/__init__.py
@@ -6,4 +6,5 @@ from ._version import __version__
 
 os.environ["TRANSFORMERS_NO_ADVISORY_WARNINGS"] = "1"
 
-from .utils import convert, generate, load, stream_generate
+from .convert import convert
+from .utils import generate, load, stream_generate
